# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat Systems API",
    "description": "Every error response carries `{\"error\": \"<message>\"}` with the HTTP status saying what kind of error it is.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/channels/{ch_id}/members": {
      "get": {
        "tags": [
          "channels"
        ],
        "operationId": "list_channel_members",
        "parameters": [
          {
            "name": "ch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataList_ChannelMember"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "channels"
        ],
        "operationId": "add_channel_member",
        "parameters": [
          {
            "name": "ch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddChannelMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelMember"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddChannelMemberRequest": {
        "type": "object",
        "required": [
          "user_id"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ChannelMember": {
        "type": "object",
        "required": [
          "channel_id",
          "user_id",
          "role",
          "notifications",
          "is_muted",
          "is_starred",
          "joined_at"
        ],
        "properties": {
          "channel_id": {
            "type": "string",
            "format": "uuid"
          },
          "is_muted": {
            "type": "boolean"
          },
          "is_starred": {
            "type": "boolean"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_read_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_read_msg": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "notifications": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ChannelRole"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ChannelRole": {
        "type": "string",
        "enum": [
          "member",
          "admin"
        ]
      },
...
      "DataList_ChannelMember": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "channel_id",
                "user_id",
                "role",
                "notifications",
                "is_muted",
                "is_starred",
                "joined_at"
              ],
              "properties": {
                "channel_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "is_muted": {
                  "type": "boolean"
                },
                "is_starred": {
                  "type": "boolean"
                },
                "joined_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "last_read_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "last_read_msg": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "notifications": {
                  "type": "string"
                },
                "role": {
                  "$ref": "#/components/schemas/ChannelRole"
                },
                "user_id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "timezone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UserPublic": {
        "type": "object",
        "required": [
          "id",
          "email",
          "timezone",
          "status",
          "is_instance_admin",
          "created_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_instance_admin": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "status_emoji": {
            "type": [
              "string",
              "null"
            ]
          },
          "status_expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "timezone": {
            "type": "string"
          }
        }
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "suspended"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Same-origin clients are authenticated by the httpOnly cookie instead."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}